refactor(desktop): extract the SSH transport into boxssh.go
The SSH flag-set fallback chain, the subprocess runners, error classification,
the shared handshake/reboot helpers and the TCP reachability probe lived in
install_str.go but are the transport for six features (install, OTA, uninstall,
true factory reset, repair-over-SSH, log export). Move them to boxssh.go so a
reader changing the installer is no longer editing the SSH stack under five
other features, and delete the dead waitForTCPPort (no call sites). Pure move,
same package, no behavior change.
